Add files via upload
Is essentially the same problem than 1774 with a higher upper bound. This solution is a bit more efficient than the used in 1774 because of the priority queue ( the time performance with the PQ is 0.032s and with the sorted array is 0.040s). The solution using the array is in the commented lines